HP, Damage from Int32 to Double

+++ TankBattle/Classes/Ammunition.cs
@@ -6,7 +6,7 @@ namespace TankBattle.Classes
 {
     abstract class Ammo
     {
-        public Int32 Damage { get; protected set; }
+        public Double Damage { get; protected set; }
         public Int32 AmmoWeight { get; protected set; }
     }
 }
